requirements.txt
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.6.0
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
requests>=2.31.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
